pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "luansilva25", email = "luan250807@gmail.com"},
]
dependencies = [
    "django>=5.0.6",
    "djangorestframework>=3.15.2",
    "django-cors-headers>=4.4.0",
    "gunicorn>=22.0.0",
    "Flask>=3.0.3",
    "waitress>=3.0.0",
]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false

[tool.pdm.scripts]
dev = 'python manage.py runserver'
migrate = 'python manage.py migrate'
make = 'python manage.py makemigrations'

